anomaly-detector/requirements.txt
flask==3.1.0
flask-socketio==5.4.1
influxdb-client==1.46.0
python-dotenv==1.0.1
gunicorn==21.2.0
eventlet==0.36.1
